# Forward declare some things to speed up compile