--- a/src/Application.hpp
+++ b/src/Application.hpp
@@ -1,7 +1,5 @@
 #pragma once
 
-#include "singletons/NativeMessaging.hpp"
-
 #include <cassert>
 #include <memory>
 
@@ -54,6 +52,7 @@ class SeventvEventAPI;
 class ILinkResolver;
 class IStreamerMode;
 class ITwitchUsers;
+class NativeMessagingServer;
 namespace pronouns {
     class Pronouns;
 }  // namespace pronouns
@@ -236,7 +235,7 @@ public:
 private:
     void initNm(const Paths &paths);
 
-    NativeMessagingServer nmServer;
+    std::unique_ptr<NativeMessagingServer> nmServer;
     Updates &updates;
 
     bool initialized{false};