--- a/src/controllers/completion/sources/EmoteSource.cpp
+++ b/src/controllers/completion/sources/EmoteSource.cpp
@@ -11,6 +11,7 @@
 #include "providers/twitch/TwitchChannel.hpp"
 #include "providers/twitch/TwitchIrcServer.hpp"
 #include "singletons/Emotes.hpp"
+#include "widgets/splits/InputCompletionItem.hpp"
 
 namespace chatterino::completion {